app/requirements.txt
pymongo
requests
dnspython
python-decouple
autopep8
psycopg2
sqlalchemy
aiohttp
aiolimiter
aioretry